Move some classes to Internal name space
