+++ src/Provider40/Query/Sql/Internal/ISqlCeExpressionVisitor.cs
@@ -1,8 +1,9 @@
 ﻿using System.Linq.Expressions;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.Query.Expressions;
+using Microsoft.EntityFrameworkCore.Query.Expressions.Internal;
 
-namespace Microsoft.EntityFrameworkCore.Query.Sql
+namespace Microsoft.EntityFrameworkCore.Query.Sql.Internal
 {
     public interface ISqlCeExpressionVisitor
     {
